Conditional validations for questions Each question needs to be validated after the 'answer' attribute has been set by the user. Each question needs to be validated according to its kind. Add #answer_provided? method that checks that answer is present and is not {} Add #validate_answer_based_on_question_kind, which enforces different validations depending on the kind of question, and run it if answer_provided?
Fix issue with stimulus service_checkbox_controller
Checking a unchecked box (not connected to an existing record) would raise an error becasue destroyInputTarget was not on the page, as it was rendered only for existing records.
Introduced a check before running the method. If destroyInputTarget exists, the method can run, if not, it doesn't and no errors are raised.